deals_backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --allow-unsafe requirements/requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.3
    # via
    #   -r requirements/requirements.in
    #   django-admin-autocomplete-filter
    #   djangorestframework
    #   drf-yasg
django-admin-autocomplete-filter==0.7.1
    # via -r requirements/requirements.in
django-admin-rangefilter==0.10.0
    # via -r requirements/requirements.in
django-query-counter==0.3.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements/requirements.in
gunicorn==21.2.0
    # via -r requirements/requirements.in
inflection==0.5.1
    # via drf-yasg
numpy==1.25.2
    # via pandas
packaging==23.1
    # via
    #   drf-yasg
    #   gunicorn
pandas==2.0.3
    # via -r requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements/requirements.in
pytz==2023.3
    # via
    #   djangorestframework
    #   drf-yasg
    #   pandas
pyyaml==6.0.1
    # via drf-yasg
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tabulate==0.9.0
    # via django-query-counter
typing-extensions==4.7.1
    # via asgiref
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via drf-yasg
